+++ test/config/aw-config.js
@@ -15,7 +15,8 @@ var repoConfig = {
         server: {
           baseDir: "./",
           routes: {
-            "/fixtures": "test/integration"
+            "/fixtures": "test/integration",
+            "/plugins": "../../plugins"
           }
         },
         scrollRestoreTechnique: "cookie"